CIC test works, genplot tweaked
CIC test is not correctly finding the parameters for the
frequency response for a given CIC configuration. The genplot
command is now producing 'genplot-out.plot' rather than what it was
producing since what it had been producing was hard to remember.

Also genplot is still a bit broken, it can't do titles.