Pretty comprehensive rework of plot
Completely refactored plot.c/plot.h to design an API that would
allow for simpler use. The goal is to get rid of the copypasta
code in the various examples and tests. This should also help
unify the "look" for all the plots from all the programs since I
would change things like fonts on one and not update all the rest.

Still to do is to get multiplot integrated/working.